Updated files for adding to cart from product page
Several new promotions have been launched for the online store, one of which has resulted in a bug. The promotion is enabled by adding the ?promo=offerN parameter to the product link. The detected bug is not going to be fixed yet, so the failing test is marked as xfail.